requirements.txt
PyQt6>=6.6.0
mss>=9.0.0
Pillow>=10.0.0
openai>=1.52.0
python-dotenv>=1.0.0
keyboard>=0.13.5
pyperclip>=1.8.0
requests>=2.31.0
duckduckgo-search>=7.0.0
json-repair>=0.28.0
